feat: add Theme Color components and move AnimatedCard to commonEffect - Added Theme Color components to the project. - Moved AnimatedCard from src/components/common/ to src/components/commonEffect/.
