+++ frontend/src/components/common/colors.js
@@ -1,0 +1,27 @@
+  // 배경 색상 정의
+  export const backgroundColor = {
+    background: '#F2F2F2' // White Smoke
+  };
+  //그림자 색상 정의
+  export const shadowColor = {
+    shadow : 'D5D3CB'
+  }
+  
+  // 로고 색상 정의
+  export const logoColors = {
+    verdunGreen: '#565903', // Verdun Green
+  };
+  
+  // 버튼 색상 정의
+  export const buttonColors = {
+    bone: '#DBC7B5',
+    mediumWood: '#A6836F',
+    santeFe: '#A67153',
+    newAmbe: '#6E3C21',
+    brownPod: '#40190B',
+  };
+  // 폰트 색상 정의
+  export const fontColor = {
+    woodBark: '#30231C'
+  }
+
