+++ frontend/src/components/commonEffect/AnimatedCard.jsx
